make interest category
